% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newDOI.R
\name{newDOI}
\alias{newDOI}
\title{Import and validate new DOI to DOI list.}
\usage{
newDOI(DOI, bib = "./inSilecoRef.bib")
}
\arguments{
\item{DOI}{character vector of DOI to import}

\item{bib}{character path to master .bib file, default is `inSilecoRef.bib`}
}
\description{
Imports and validates new entries for a DOI list
}
\examples{
# Example 1:
newDOI(DOI = c('10.1016/j.tree.2016.10.011','10.1890/130230','10.1515/9781400881376'))
}
\keyword{DOI}
