% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_cran.R
\name{url_cran}
\alias{url_cran}
\alias{url_bioc}
\alias{url_gh}
\title{URL for a specific package}
\usage{
url_cran(pkg = "")

url_bioc(pkg = "")

url_gh(pkg = "", user = "cran")
}
\arguments{
\item{pkg}{a package name.}

\item{user}{a character string of the GitHub repository.}
}
\value{
A link to the CRAN's canonical URL for a given package formatted in Markdown.
Note that is \code{pkg = ""} a link to the CRAN is returned.
}
\description{
Get an URL for a specific package (Markdown style).
}
\section{Functions}{
\itemize{
\item \code{url_bioc}: A link to the Bioconductor's short URL for a given
package formatted in Markdown. Note that is \code{pkg = ""} a link to Bioconductor
is returned.

\item \code{url_gh}: A link to the GitHub repository for a given package.
If \code{pkg = ""} a link to the user account in returned and if \code{pkg = ""} and
}}

\examples{
url_cran("")
url_cran("ps")
url_gh("inSilecoRef", "inSileco")
url_bioc("CRImage")
}
