% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_pkg_cran.R
\name{url_pkg_cran}
\alias{url_pkg_cran}
\alias{url_pkg_gh}
\title{URL for a specific package}
\usage{
url_pkg_cran(pkg)

url_pkg_gh(pkg, user = "cran")
}
\arguments{
\item{pkg}{a package name.}

\item{user}{a character string of the GitHub repository.}
}
\value{
A link to the CRAN's canonical URL for a given formatted in Markdown.
}
\description{
Get an URL for a specific package (Markdown style).
}
\section{Functions}{
\itemize{
\item \code{url_pkg_gh}: A link to the GitHub repository for a given package.
}}

\examples{
url_pkg_cran("ps")
url_pkg_gh("inSilecoRef", "inSileco")
}
